small renaming input -> inputs
