+++ frontend/src/shared/ui/controls/inputs/PasswordInput/PasswordInput.tsx
@@ -3,8 +3,8 @@ import { FC, useState } from "react";
 import { SrOnlyLabel } from "@/shared/styles/globalStyledComponents";
 import { EyeNoneIcon, EyeOpenIcon } from "@radix-ui/react-icons";
 
-import * as S from "../input.styles";
-import { IInputProps } from "../input.types";
+import * as S from "../inputs.styles";
+import { IInputProps } from "../inputs.types";
 
 export const PasswordInput: FC<IInputProps> = ({ id, label, required, placeholder, error, field }) => {
   const [isPasswordRevealed, setIsPasswordRevealed] = useState<boolean>(false);
